# portal/api/tool/v1/tool_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.0
// source: tool/v1/tool.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationToolCreateTool = "/tool.v1.Tool/CreateTool"
const OperationToolDeleteTool = "/tool.v1.Tool/DeleteTool"
const OperationToolGetTool = "/tool.v1.Tool/GetTool"
const OperationToolListTools = "/tool.v1.Tool/ListTools"
const OperationToolUpdateTool = "/tool.v1.Tool/UpdateTool"

type ToolHTTPServer interface {
	// CreateTool 创建工具
	CreateTool(context.Context, *CreateToolRequest) (*ToolReply, error)
	// DeleteTool 删除工具（若被绑定返回 409）
	DeleteTool(context.Context, *DeleteToolRequest) (*DeleteToolReply, error)
	// GetTool 获取单个工具
	GetTool(context.Context, *GetToolRequest) (*ToolReply, error)
	// ListTools 工具列表（分页、筛选）
	ListTools(context.Context, *ListToolsRequest) (*ListToolsReply, error)
	// UpdateTool 更新工具
	UpdateTool(context.Context, *UpdateToolRequest) (*ToolReply, error)
}

func RegisterToolHTTPServer(s *http.Server, srv ToolHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/tools", _Tool_CreateTool0_HTTP_Handler(srv))
	r.GET("/api/v1/tools/{id}", _Tool_GetTool0_HTTP_Handler(srv))
	r.GET("/api/v1/tools", _Tool_ListTools0_HTTP_Handler(srv))
	r.PUT("/api/v1/tools/{id}", _Tool_UpdateTool0_HTTP_Handler(srv))
	r.DELETE("/api/v1/tools/{id}", _Tool_DeleteTool0_HTTP_Handler(srv))
}

func _Tool_CreateTool0_HTTP_Handler(srv ToolHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateToolRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationToolCreateTool)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateTool(ctx, req.(*CreateToolRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ToolReply)
		return ctx.Result(200, reply)
	}
}

func _Tool_GetTool0_HTTP_Handler(srv ToolHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetToolRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationToolGetTool)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetTool(ctx, req.(*GetToolRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ToolReply)
		return ctx.Result(200, reply)
	}
}

func _Tool_ListTools0_HTTP_Handler(srv ToolHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListToolsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationToolListTools)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListTools(ctx, req.(*ListToolsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListToolsReply)
		return ctx.Result(200, reply)
	}
}

func _Tool_UpdateTool0_HTTP_Handler(srv ToolHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateToolRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationToolUpdateTool)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateTool(ctx, req.(*UpdateToolRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ToolReply)
		return ctx.Result(200, reply)
	}
}

func _Tool_DeleteTool0_HTTP_Handler(srv ToolHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteToolRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationToolDeleteTool)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteTool(ctx, req.(*DeleteToolRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteToolReply)
		return ctx.Result(200, reply)
	}
}

type ToolHTTPClient interface {
	// CreateTool 创建工具
	CreateTool(ctx context.Context, req *CreateToolRequest, opts ...http.CallOption) (rsp *ToolReply, err error)
	// DeleteTool 删除工具（若被绑定返回 409）
	DeleteTool(ctx context.Context, req *DeleteToolRequest, opts ...http.CallOption) (rsp *DeleteToolReply, err error)
	// GetTool 获取单个工具
	GetTool(ctx context.Context, req *GetToolRequest, opts ...http.CallOption) (rsp *ToolReply, err error)
	// ListTools 工具列表（分页、筛选）
	ListTools(ctx context.Context, req *ListToolsRequest, opts ...http.CallOption) (rsp *ListToolsReply, err error)
	// UpdateTool 更新工具
	UpdateTool(ctx context.Context, req *UpdateToolRequest, opts ...http.CallOption) (rsp *ToolReply, err error)
}

type ToolHTTPClientImpl struct {
	cc *http.Client
}

func NewToolHTTPClient(client *http.Client) ToolHTTPClient {
	return &ToolHTTPClientImpl{client}
}

// CreateTool 创建工具
func (c *ToolHTTPClientImpl) CreateTool(ctx context.Context, in *CreateToolRequest, opts ...http.CallOption) (*ToolReply, error) {
	var out ToolReply
	pattern := "/api/v1/tools"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationToolCreateTool))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteTool 删除工具（若被绑定返回 409）
func (c *ToolHTTPClientImpl) DeleteTool(ctx context.Context, in *DeleteToolRequest, opts ...http.CallOption) (*DeleteToolReply, error) {
	var out DeleteToolReply
	pattern := "/api/v1/tools/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationToolDeleteTool))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetTool 获取单个工具
func (c *ToolHTTPClientImpl) GetTool(ctx context.Context, in *GetToolRequest, opts ...http.CallOption) (*ToolReply, error) {
	var out ToolReply
	pattern := "/api/v1/tools/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationToolGetTool))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListTools 工具列表（分页、筛选）
func (c *ToolHTTPClientImpl) ListTools(ctx context.Context, in *ListToolsRequest, opts ...http.CallOption) (*ListToolsReply, error) {
	var out ListToolsReply
	pattern := "/api/v1/tools"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationToolListTools))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateTool 更新工具
func (c *ToolHTTPClientImpl) UpdateTool(ctx context.Context, in *UpdateToolRequest, opts ...http.CallOption) (*ToolReply, error) {
	var out ToolReply
	pattern := "/api/v1/tools/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationToolUpdateTool))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}