# portal/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/agents/{id}/tools:
        post:
            tags:
                - Agent
            description: 绑定工具
            operationId: Agent_BindTools
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/agent.v1.BindToolsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/agent.v1.BindToolsReply'
        delete:
            tags:
                - Agent
            description: 解绑工具
            operationId: Agent_UnbindTools
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: toolIds
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/agent.v1.UnbindToolsReply'
...
    /api/v1/tools:
        get:
            tags:
                - Tool
            description: 工具列表（分页、筛选）
            operationId: Tool_ListTools
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: name
                  in: query
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tool.v1.ListToolsReply'
        post:
            tags:
                - Tool
            description: 创建工具
            operationId: Tool_CreateTool
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/tool.v1.CreateToolRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tool.v1.ToolReply'
    /api/v1/tools/{id}:
        get:
            tags:
                - Tool
            description: 获取单个工具
            operationId: Tool_GetTool
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tool.v1.ToolReply'
        put:
            tags:
                - Tool
            description: 更新工具
            operationId: Tool_UpdateTool
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/tool.v1.UpdateToolRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tool.v1.ToolReply'
        delete:
            tags:
                - Tool
            description: 删除工具（若被绑定返回 409）
            operationId: Tool_DeleteTool
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tool.v1.DeleteToolReply'
...
components:
    schemas:
...
        agent.v1.BindToolsReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
        agent.v1.BindToolsRequest:
            type: object
            properties:
                id:
                    type: string
                toolIds:
                    type: array
                    items:
                        type: string
...
        agent.v1.UnbindToolsReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
...
        api.common.BaseResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                reason:
                    type: string
            description: 统一响应基类，所有 Response/Reply 需嵌入
...
        tool.v1.CreateToolRequest:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                type:
                    type: string
                config:
                    $ref: '#/components/schemas/tool.v1.ToolConfig'
        tool.v1.DatasourceConfig:
            type: object
            properties:
                id:
                    type: string
                type:
                    type: string
                dsn:
                    type: string
                host:
                    type: string
                port:
                    type: integer
                    format: int32
                user:
                    type: string
                password:
                    type: string
                dbname:
                    type: string
                readOnly:
                    type: boolean
            description: DatasourceConfig 数据源配置，与 framework/datasource.Config 对齐
        tool.v1.DeleteToolReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
        tool.v1.ListToolsReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/tool.v1.ToolReply'
                total:
                    type: integer
                    format: int32
        tool.v1.McpConfig:
            type: object
            properties:
                endpoint:
                    type: string
                id:
                    type: string
                backend:
                    type: string
            description: McpConfig MCP 工具运行时配置，与 framework/tool.McpConfig 对齐
        tool.v1.RCAConfig:
            type: object
            properties:
                funcPath:
                    type: string
                roots:
                    type: array
                    items:
                        type: string
                queryUrl:
                    type: string
                datasourceId:
                    type: string
                defaultIndex:
                    type: string
                traceIdField:
                    type: string
            description: RCAConfig 根因分析工具配置。func_path 指定子工具:rca_code | jaeger_trace | es_log_query。
        tool.v1.ToolConfig:
            type: object
            properties:
                funcPath:
                    type: string
                parameters:
                    type: object
                async:
                    type: boolean
                mcpServerId:
                    type: string
                    description: MCP 配置（兼容旧字段，优先使用 mcp 嵌套）
                mcpEndpoint:
                    type: string
                mcpBackend:
                    type: string
                timeoutSec:
                    type: integer
                    format: int32
                mcp:
                    $ref: '#/components/schemas/tool.v1.McpConfig'
                datasource:
                    $ref: '#/components/schemas/tool.v1.DatasourceConfig'
                rca:
                    $ref: '#/components/schemas/tool.v1.RCAConfig'
            description: 工具配置：内置用 func_path/parameters，MCP 用 mcp 或 mcp_* 字段，数据源用 datasource
        tool.v1.ToolReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
                id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                type:
                    type: string
                config:
                    $ref: '#/components/schemas/tool.v1.ToolConfig'
                createdAt:
                    type: string
                updatedAt:
                    type: string
        tool.v1.UpdateToolRequest:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                type:
                    type: string
                config:
                    $ref: '#/components/schemas/tool.v1.ToolConfig'
...